requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
edge-tts==7.2.7
frozenlist==1.8.0
google-genai==1.63.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
moviepy==2.2.1
multidict==6.7.1
numpy==2.4.2
pillow==11.3.0
proglog==0.1.12
propcache==0.4.1
python-dotenv==1.2.1
requests==2.32.5
tabulate==0.9.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0
